EC10 PoC
When calling minio server with a single path
an FSObjectLayer is usually created.

When calling minio server with multiple paths
usually an ErasureServerPool (EC) is created.

S3 Versioning and ObjectLock features are bound
to the EC scenario so it is not available
when you only specify a single path.

This commit enables Versioning and Object-Locking
by allowing  an EC10 (1 data, 0 parity) algo
instead of the FSObjectLayer.