REVERT THIS COMMIT!!! Contains localized testing scripts.
These scripts are only usable in a specific environment. I could easily
make them more generic, but I don't see the point. Mint will be run from
a docker container in most cases anyway and these are easy enough to
reproduce in case we need to adjust some of the test packages.

Sadly I was unable to reproduce the issue with the minio-go tests
outside of docker, although I'm sure that the underlying cause is the
same. I'll try to get my code changes into the docker container to
verify them.