REVERT THIS COMMIT!!! Contains docker files to reproduce port issue!
I wrote some docker files to reproduce the bug where any Mint tests
using a Go SDK are unable to send valid presigned requests.

The cause for this is the usage of a default port, either 443 or 80. The
default port is stripped from the calculated signature, but kept in the
service URL and Host header, which makes MinIO calculate a different
signature.

The issue can easily be reproduced with the files at hand, so they
should be included in a bug report.

To test this with a TLS encrypted MinIO instance, we need to generate at
least a self-signed certificate in the docker container. Sadly the
default MinIO image does not include

- openssl,
- wget,
- tar,
- yum,
- apt or
- any other way to install software from a package repository.

That's what the minio `Dockerfile` is for. As a base image I chose ubuntu, but
it could have been anything else.

The mint `Dockerfile` allows me to easily test my changes. We don't need
to include this in the bug report. The MinIO guys will have their own
tooling to build a new image.

I also added bash scripts to help with building the docker images. In
case of the mint dockerfile we need to copy a bunch of files around and
build the binaries with go before we copy them to the image.